General issue patches: - Comments now annotated with #item# - Added list expression. - Numbers are now floats rather then ints.

+++ src/base/Interpreter.cpp
@@ -214,7 +214,7 @@ void Interpreter::interpret(vector<Token> &tokens, int limit, const string &endT
 
         switch (token.type) {
             case NUMBER: {
-                auto c = std::make_unique<NumberExpression>(std::stoi(token.tokenId));
+                auto c = std::make_unique<NumberExpression>(std::stof(token.tokenId));
                 c->lineNumber = token.lineNumber;
                 abstractSyntaxTree.push_back(std::move(c));
                 break;
